backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.3
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.12.1
annotated-types==0.7.0
anthropic==0.18.0
anyio==3.7.1
attrs==25.3.0
bcrypt==4.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.2
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.104.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.23.4
google-auth-httplib2
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
googlemaps==4.10.0
greenlet==3.2.1
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.31.4
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
Mako==1.3.10
MarkupSafe==3.0.2
motor==3.7.1
multidict==6.4.3
numpy==2.2.6
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pluggy==1.6.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.10.1
PyJWT==2.10.1
pymongo==4.13.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.22
starlette==0.27.0
stripe==12.1.0
tokenizers==0.21.1
tqdm==4.67.1
twilio==8.10.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.24.0
yarl==1.20.0